+++ src/luxos/cli/shared.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import argparse
+import inspect
 import sys
 import types
 from typing import Callable
@@ -13,6 +14,15 @@ else:
     ArgsCallback = Callable
 
 
+def check_default_constructor(klass: type):
+    signature = inspect.signature(klass.__init__)  # type: ignore[misc]
+    for name, value in signature.parameters.items():
+        if name in {"self", "args", "kwargs"}:
+            continue
+        if value.default is inspect.Signature.empty:
+            raise RuntimeError(f"the {klass}() cannot be called without arguments")
+
+
 # The luxor base parser
 class LuxosParserBase(argparse.ArgumentParser):
     def __init__(self, modules: list[types.ModuleType], *args, **kwargs):
@@ -21,12 +31,17 @@ class LuxosParserBase(argparse.ArgumentParser):
         self.callbacks: list[ArgsCallback | None] = []
 
     def add_argument(self, *args, **kwargs):
-        try:
-            if issubclass(kwargs.get("type"), ArgumentTypeBase):
-                kwargs["default"] = kwargs.get("type")(kwargs.get("default"))
-                kwargs["type"] = kwargs["default"]
-        except TypeError:
-            pass
+        typ = kwargs.get("type")
+        obj = None
+        if isinstance(typ, type) and issubclass(typ, ArgumentTypeBase):
+            check_default_constructor(typ)
+            obj = typ()
+        if isinstance(typ, ArgumentTypeBase):
+            obj = typ
+        if obj is not None:
+            obj.default = kwargs.get("default")
+            kwargs["default"] = obj
+            kwargs["type"] = obj
         super().add_argument(*args, **kwargs)
 
 
@@ -34,10 +49,19 @@ class ArgumentTypeBase:
     class _NA:
         pass
 
-    def __init__(self, default=_NA):
-        self.default = default
-        if default is not ArgumentTypeBase._NA:
-            self.default = self._validate(default)
+    def __init__(self, *args, **kwargs):
+        self.args = args
+        self.kwargs = kwargs
+        self._default = self._NA
+
+    @property
+    def default(self):
+        return self._default
+
+    @default.setter
+    def default(self, value):
+        self._default = self._validate(value)
+        return self._default
 
     def __call__(self, txt):
         self._value = None